ocr-service/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# OCR Engines
paddleocr==2.8.1
easyocr==1.7.2
pytesseract==0.3.13

# Image Processing
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
scikit-image==0.24.0
imutils==0.5.4

# Redis for queue and caching
redis==5.2.0
rq==1.16.2
celery[redis]==5.4.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2
regex==2024.9.11

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==5.0.0
